pyproject.toml
[project]
name = "sax-buddy-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "firebase-admin>=6.9.0",
    "pydantic[email]>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "uvicorn>=0.35.0",
    "python-multipart>=0.0.5",
    "python-dotenv>=1.1.1",
]

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
